fix: Ensure screen API requests receive 404 for nonexistent screen IDs
